Proper Container Shutdown Procedure
Containers are now properly shut down, via docker-compose and a
follow-up docker stop on any stragglers, to ensure that if a process had
a volume open, that it would relinquish that control after a monit stop.

The side-effect of this is that if you stop and (re)start the `docker`
monit job manually, on-box, the `post-start` script doesn't run and you
don't get back your compose'd containers.  A simple fix is to manually
trigger the post-start yourself.

If, however, you `bosh stop` / `bosh start` (from off-box), everything
works, since BOSH knows to trigger the post-start for you.

Fixes #2