Treat YAML files as ... YAML files
Normally, one ought to specify a YAML file for ttar (via the `files`
property) as a string of YAML, embedded in a BOSH YAML manifest,
like so:

    properties:
      files:
        my.yaml: |
          ---
          embedded: yaml

However, it makes sense, especially with Spruce's `(( vault ... ))` and
`(( file ... ))` operators, or `((credhub))` notation, to want to
continue the YAML love into the file, and have the BOSH release render
it into a string for us, i.e.

    properties:
      files:
        secrets.yml:
          foo: (( vault "secret/foo" ))

This now works, as expected.  Any non-string file contents are
explicitly turned into YAML via a `to_yaml` call in the ERB.

Fixes #5