Don't let bad docker-compose.yml ruin a good stop
BOSH used to get angry during a `monit stop` of the `docker` job if the
docker-compose.yml recipe was bad (mismatched version, was actually a
quoted string, bad YAML contents, etc.).

Since this happens quite a lot, especially when I'm involved, I finally
fixed this so that a compose down wasn't the end of the world, and we
could force-kill the containers and be done with it.