Allow string recipe: values
My original desire was to be able to spruce merge _through_ a recipe in
the BOSH manifest.  However, as we saw in #11, there are times where we
want the recipe to come from outside, via something like `(( file ... ))`

Now, if `recipe` is a string, we pass it through as-is.  Who are we to
be more clever than the operations folks doing deployments?