Support files as an array, for go-patch users
If you've made life decisions that have led you down the path of
go-patch operator files, you'll be pleased to know that we are
particularly charitable.  You can specify the `files` directive as an
array of maps, which will each be merged in-line during template render.

This is easier to do with Spruce and tools like it, but sometimes you're
stuck with the decisions of the past.